Add LTI Support and improve UI
Update the main controller and template so that it can display adequate
information depending on the user being on an LTI launch or direct access.

The global UI of the app has been improve (at least for the public part).

Deployment configurations have been added to deploy the app on the
production and staging environments.